Split OpenSSL backend into separate files (hmac/cmac)
